+++ grumpy/cli/about.py
@@ -1,4 +1,5 @@
 from cleo import Command
+from rich.table import Table
 from ..commands import ALL_CHECKS
 from ..commands import FINAL_MESSAGE
 from .cat import cat, random_phrase
@@ -11,9 +12,13 @@ class AboutCommand(Command):
 
     about
         {command : What are you grumpy about?}
+        {--topic=? : For asking about a specific topic}
     """
+    
 
     def handle(self):
+        if self.argument('command') == "what":
+            return self._handle_what()
         command = 'check_' + self.argument('command')
         ALL_CHECKS[command]()
         if FINAL_MESSAGE['issues'] == 0:
@@ -25,3 +30,14 @@ class AboutCommand(Command):
             )
         else:
             console.print(random_phrase())
+    
+    def _handle_what(self):
+        what_option = self.option('topic')
+        
+        available_commands = [s.replace("check_", "") for s in ALL_CHECKS.keys()]
+        table = Table(show_header=True)
+        table.add_column("Command")
+        table.add_column("Description")
+        for com in available_commands: 
+            table.add_row(com, "")
+        console.print(table)